+++ addon/templates/components/medium-editor.hbs
@@ -1,0 +1,1 @@
+TEst
