+++ ember-cli-build.js
@@ -6,6 +6,9 @@ module.exports = function(defaults) {
   let app = new EmberAddon(defaults, {
     'ember-cli-babel': {
       includePolyfill: true
+    },
+    vendorFiles: {
+      'jquery.js': null
     }
   });
