+++ package.json
@@ -28,6 +28,7 @@
     "broccoli-merge-trees": "^2.0.0",
     "broccoli-stew": "^1.5.0",
     "ember-cli-babel": "^6.6.0",
+    "ember-cli-htmlbars": "^2.0.3",
     "medium-editor": "^5.23.0"
   },
   "devDependencies": {
@@ -36,15 +37,16 @@
     "ember-cli": "~2.18.2",
     "ember-cli-dependency-checker": "^2.0.0",
     "ember-cli-eslint": "4",
-    "ember-cli-fastboot": "1.0.0-rc.2",
-    "ember-cli-htmlbars": "^2.0.1",
+    "ember-cli-fastboot": "^1.1.3",
     "ember-cli-htmlbars-inline-precompile": "^1.0.0",
     "ember-cli-inject-live-reload": "^1.4.1",
     "ember-cli-qunit": "^4.1.1",
     "ember-cli-shims": "^1.2.0",
     "ember-cli-sri": "^2.1.0",
     "ember-cli-uglify": "^2.0.0",
+    "ember-cli-update": "^0.14.0",
     "ember-cli-yuidoc": "^0.8.8",
+    "ember-composability-tools": "^0.0.9",
     "ember-disable-prototype-extensions": "^1.1.2",
     "ember-export-application-global": "^2.0.0",
     "ember-load-initializers": "^1.0.0",
